pyproject.toml
[tool.poetry]
name = "flask_realworld"
version = "0.1"
description = "flask realworld project"
authors = ["hamedsh <sheykhlou.h@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
Flask = "^2.0.1"
Flask-Cors = "^3.0.10"
alembic = "^1.6.5"
Flask-Migrate = "^3.0.1"
Flask-Bcrypt = "^0.7.1"
Flask-Caching = "^1.10.1"
Flask-SQLAlchemy = "^2.5.1"
unicode-slugify = "^0.1.3"
Flask-JWT-Extended = "^4.2.3"
marshmallow = "^3.12.1"
flask-apispec = "^0.11.0"

[tool.poetry.dev-dependencies]
mypy = "^0.902"
flakehell = "^0.9.0"
flake8-docstrings = "^1.6.0"
flake8-commas = "^2.0.0"
flake8-quotes = "^3.2.0"
flake8-bandit = "^2.1.2"
black = "^21.6b0"
flake8-bugbear = "^21.4.3"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.5.0"
flake8-eradicate = "^1.0.0"
flake8-isort = "^4.0.0"
flake8-mutable = "^1.2.0"
flake8-pytest-style = "^1.4.2"
pep8-naming = "^0.11.1"
tox = "^3.23.1"
pytest = "^6.2.4"
sqlalchemy-stubs = "^0.4"
flask-sqlalchemy-stubs = "^0.2"
factory-boy = "^3.2.0"
WebTest = "^2.0.35"
pycrunch-engine = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# -----------------------------------------------flakehell
[tool.flakehell]
extended_default_ignore=[]
exclude = ["README.rst", "README.md"]
format = "colored"
max_line_length = 88
show_source = true
whitelist = "../../allowlist.txt"

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
#flake8-darglint = ["+*"]
flake8-docstrings = ["+*", "-D100"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
#flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*"]
pyflakes = ["+*"]
pylint = ["+*"]

#-----------------------------------------------black
[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
